// env.d.ts
type NivaSize = {
  /** 宽 */
  width: number;
  /** 高 */
  height: number;
};

/** 位置 */
type NivaPosition = {
  /** 距离坐标原点横轴距离 */
  x: number;
  /** 距离坐标原点纵轴距离 */
  y: number;
};

/** 窗口根菜单 */
interface WindowRootMenu {
  /** 菜单项的名称 */
  label: string;
  /** 是否启用 */
  enabled?: boolean;
  /** 子菜单项 */
  children: MenuOptions;
}

/** 窗口根菜单集合 */
type WindowMenuOptions = Array<WindowRootMenu>;

/** 窗口选项 */
interface NivaWindowOptions {
  /** 应用程序入口文件路径 */
  entry?: string;
  /** 是否启用开发者工具 */
  devtools?: boolean;

  /** 窗口标题 */
  title?: string;
  /** 窗口图标 */
  icon?: string;
  /** 窗口主题 */
  theme?: string;
  /** 窗口大小 */
  size?: NivaSize;
  /** 窗口最小尺寸 */
  minSize?: NivaSize;
  /** 窗口最大尺寸 */
  maxSize?: NivaSize;

  /** 窗口位置 */
  position?: NivaPosition;

  /** 是否可调整大小 */
  resizable?: boolean;
  /** 是否可最小化 */
  minimizable?: boolean;
  /** 是否可最大化 */
  maximizable?: boolean;
  /** f */
  closable?: boolean;

  /** 是否全屏 */
  fullscreen?: boolean;
  /** 是否最大化 */
  maximized?: boolean;
  /** 是否可见 */
  visible?: boolean;
  /** 是否透明 */
  transparent?: boolean;
  /** 是否显示窗口装饰(标题栏) */
  decorations?: boolean;

  /** 是否始终置顶 */
  alwaysOnTop?: boolean;
  /** 是否始终置底 */
  alwaysOnBottom?: boolean;
  /** 是否在多个工作区显示 */
  visibleOnAllWorkspaces?: boolean;

  /** 是否聚焦 */
  focused?: boolean;
  /** 是否启用内容保护 */
  contentProtection?: boolean;

  // windows extra
  /** 父窗口 ID */
  parentWindow?: number;
  /** 拥有者窗口 ID */
  ownerWindow?: number;
  /** 任务栏图标 */
  taskbarIcon?: string;
  /** 在任务栏中是否显示 */
  skipTaskbar?: boolean;
  /** 是否显示窗口阴影 */
  undecoratedShadow?: boolean;

  /** 窗口菜单选项 */
  menu?: WindowMenuOptions;
}